Crate sentinel

Sentinel cryptographically confirms the origin of a claim in a decentralised network.

A claim has to implement Claimable in order to be verifiable and mergeable. A request has to implement GetSigningKeys such that Sentinel can aqcuire the necessary public signing keys. The request is passed immutably through sentinel and is used as a key to group claims and corresponding keys. When sentinel resolves a threshold on verified and merged message, it returns the requestor key and the merged claim. Claimant names and associated signatures are discarded after succesful resolution, as such abstracting the original request into a resolved claim.

The keys_threshold specifies a minimal threshold on the number of independent mentions of a single public signing key needed to consider it for verifying a claim. This threshold can be one or higher. The claims_threshold specifies a minimal threshold on the number of verified claims before sentinel will attempt to merge these verified claims.

Structs

Sentinel
Sentinel is templated on an immutable Request type, a mergeable Claim type. It further takes a Name type to identify claimants. Signature and PublicSignKey type are auxiliary types to handle a user-chosen cryptographic signing scheme.

Traits

Claimable
The Claim type needs to implement this Claimable trait. Sentinel will call verify on the claim with the associated signature and the public signing key that should verify it. Sentinel independently retrieves these public signing keys through the trait get_signing_keys on the original request. When the threshold is reached on a

GetSigningKeys
The Request type needs to implement this GetSigningKey trait. Sentinel will call get_signing_keys() the first time it receives a request for which it does not yet have any associated keys.