Module sentinel::sentinel

Sentinel provides a consensus mechanism on all content messages. The claims made must be identical and cryptographically signed. Sentinel cryptographically confirms the origin of a claim in a decentralised network.

A claim has to implement Claimable in order to be verifiable and mergeable. A request has to implement GetSigningKeys such that Sentinel can aqcuire the necessary public signing keys. The request is passed immutably through sentinel and is used as a key to group claims and corresponding keys. When sentinel resolves a threshold on verified and merged message, it returns the requestor key and the merged claim. Claimant names and associated signatures are discarded after succesful resolution, as such abstracting the original request into a resolved claim.

The keys_threshold specifies a minimal threshold on the number of independent mentions of a single public signing key needed to consider it for verifying a claim. This threshold can be one or higher. The claims_threshold specifies a minimal threshold on the number of verified claims before sentinel will attempt to merge these verified claims.

Structs

Sentinel
Sentinel is templated on an immutable Request type, a mergeable Claim type. It further takes a Name type to identify claimants. Signature and PublicSignKey type are auxiliary types to handle a user-chosen cryptographic signing scheme.

Enums

AddResult

Traits

Source