Struct sentinel::sentinel::Sentinel

pub struct Sentinel<Request, Name> where Request: Eq + PartialOrd + Ord + Clone + Source<Name>, Name: Eq + PartialOrd + Ord + Clone {
    // some fields omitted
}

Sentinel is templated on an immutable Request type, a mergeable Claim type. It further takes a Name type to identify claimants. Signature and PublicSignKey type are auxiliary types to handle a user-chosen cryptographic signing scheme.

Methods

impl<Request, Name> Sentinel<Request, Name> where Request: Eq + PartialOrd + Ord + Clone + Source<Name>, Name: Eq + PartialOrd + Ord + Clone

fn new(claim_threshold: usize, keys_threshold: usize) -> Sentinel<Request, Name>

This creates a new sentinel that will collect a minimal claim_threshold number of verified claims before attempting to merge these claims. To obtain a verified claim Sentinel needs to have received a matching public signing key. Each such a public signing key needs keys_threshold confirmations for it to be considered valid and used for verifying the signature of the corresponding claim.

fn add_claim(&mut self, request: Request, claimant: Name, signature: Signature, claim: SerialisedClaim) -> Option<AddResult<Request, Name>>

This adds a new claim for the provided request. The claimant name and the signature provided will be used to verify the claim with the keys that are independently retrieved. When an added claim leads to the resolution of the request, the request and the claim are returned. All resolved claims have to be identical. Otherwise None is returned.

Possible results are: * Some(AddResult::Resolved(request, serialised_claim)): indicating that the claim has been successfully resolved. * Some(AddResult::RequestKeys(target)): indicating that the caller should request public keys from the group surrounding the target. * None: indicating that no resolve was possible yet.

fn add_keys(&mut self, request: Request, sender: Name, keys: Vec<(Name, PublicKey)>) -> Option<(Request, SerialisedClaim)>

This adds a new set of public_signing_keys for the provided request. If the request is not known yet by sentinel, the added keys will be ignored. When the added set of keys leads to the resolution of the request, the request and the verified and merged claim is returned. Otherwise None is returned.