Struct sentinel::Sentinel

pub struct Sentinel<Request, Claim, Name, Signature, PublicSignKey> where Request: GetSigningKeys<Name> + PartialOrd + Ord + Clone, Claim: Clone + Claimable<Name, Signature, PublicSignKey> + Encodable + Decodable, Name: Eq + PartialOrd + Ord + Clone, Signature: Clone, PublicSignKey: Clone {
    // some fields omitted
}

Sentinel is templated on an immutable Request type, a mergeable Claim type. It further takes a Name type to identify claimants. Signature and PublicSignKey type are auxiliary types to handle a user-chosen cryptographic signing scheme.

Methods

impl<Request, Claim, Name, Signature, PublicSignKey> Sentinel<Request, Claim, Name, Signature, PublicSignKey> where Request: GetSigningKeys<Name> + PartialOrd + Ord + Clone, Claim: Clone + Claimable<Name, Signature, PublicSignKey> + Encodable + Decodable, Name: Eq + PartialOrd + Ord + Clone, Signature: Clone, PublicSignKey: Clone

fn new(claim_threshold: usize, keys_threshold: usize) -> Sentinel<Request, Claim, Name, Signature, PublicSignKey>

This creates a new sentinel that will collect a minimal claim_threshold number of verified claims before attempting to merge these claims. To obtain a verified claim Sentinel needs to have received a matching public signing key. Each such a public signing key needs keys_threshold confirmations for it to be considered valid and used for verifying the signature of the corresponding claim.

fn add_claim(&mut self, request: Request, claimant: Name, signature: Signature, claim: Claim) -> Option<(Request, Claim)>

This adds a new claim for the provided request. The name and the signature provided will be used to validate the claim with the keys that are independently retrieved. When an added claim leads to the resolution of the request, the request and the verified and merged claim is returned. Otherwise None is returned.

fn add_keys(&mut self, request: Request, keys: Vec<(Name, PublicSignKey)>) -> Option<(Request, Claim)>

This adds a new set of public_signing_keys for the provided request. If the request is not known yet by sentinel, the added keys will be ignored. When the added set of keys leads to the resolution of the request, the request and the verified and merged claim is returned. Otherwise None is returned.