Trait sentinel::Claimable

pub trait Claimable<Name, Signature, PublicSignKey> {
    fn verify(signature: Signature, public_key: PublicSignKey) -> bool;
    fn merge<'a, I>(xs: I) -> Option<Self> where I: Iterator<Item=&'a Self>;
}

The Claim type needs to implement this Claimable trait. Sentinel will call verify on the claim with the associated signature and the public signing key that should verify it. Sentinel independently retrieves these public signing keys through the trait get_signing_keys on the original request. When the threshold is reached on a

Required Methods

fn verify(signature: Signature, public_key: PublicSignKey) -> bool

fn merge<'a, I>(xs: I) -> Option<Self> where I: Iterator<Item=&'a Self>